Semi-working round-trip.
First, I run this WCF.Config.Helper.exe tool on my Mac.  This will
download the WSDL from my Windows box and write the config as
config.xml and config.xsd.

Then, I copy these files to a directory that's shared via apache
and run the MonoTouch test app on my iPad.  This downloads the
config.xml and config.xsd from my mac, reads the endpoint configuration
from it and attempts to connect to the server.